Add `model_options` to settings.json (#9)
This patch extends the configuration capabilities of the fm_agent.

Until now, loading the model parameters from the configuration file was
sufficient, but not all params can be set this way. The `--quantum=N` is
one example. We need to decrease this value so the timing of the model
better matches real hardware.

Signed-off-by: Filip Jagodzinski <filip.jagodzinski@arm.com>